fix dependencies issues
